Refactor: Modularize ContentCollector functionality
Broke down the monolithic ContentCollector class into smaller, more focused modules:

- ContentManager: Handles loading, saving, and basic CRUD for articles, interests, and chat highlights.
- ArticleGenerator: Responsible for generating summary articles from interests and chat highlights.
- MagazineGenerator: Manages the EPUB generation process, utilizing ContentManager and ArticleGenerator.
- Cli: Contains all command-line interface logic, including interactive prompts and argument parsing.
- TemplateManager: Provides a function to create a sample `magazine-content.json` file.

The main `content_collector.js` script now serves as a simple entry point, delegating tasks to these new modules.

Updated ContentManager to default to `out/magazine-content.json` to align with the template creation path and ensure consistent content handling.
